requirements.txt
streamlit
PyPDF2
langchain-Community
requests
python-dotenv